Support supplying a script to jk transform
This implements the `-m` (resolve as a module) flag, and the no flag
(resolve as a filesystem path) case.

This is similar to `-c` (literal JavaScript), except that the supplied
argument is used as an import, rather than the body of a function. I
decided it would be easier to use two different templates --
internal/transform-module.js and internal/transform-exec.js -- rather
than one. But this approach does mean the core bit needs to be in a
public module (transform.js).

Though I'd factored out the code for finding out an appropriate base
path for resolving things (establishScriptDir), it turns out in this
case the appropriate base path is the current directory. As with `jk
generate`, the path or module is imported from a _temporary module_,
and unless its own path is the current directory, the resolution won't
work.
