+++ std/internal/transform-exec.js
@@ -1,0 +1,8 @@
+import transform from '@jkcfg/std/transform';
+
+const makeTransformFn = new Function(`
+  return (%s);
+`);
+
+const fn = makeTransformFn();
+transform(fn);
