+++ tests/test-transform-files/addone.js
@@ -1,0 +1,1 @@
+export default v => v + 1;
