+++ tests/test-transform-script.js.cmd
@@ -1,0 +1,1 @@
+jk transform --stdout ./test-transform-files/addone.js ./test-transform-files/one.json
