+++ tests/test-transform-script.js.expected
@@ -1,0 +1,1 @@
+2
